+++ CHANGELOG.md
@@ -40,6 +40,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Java 19
 - Template in request body
 - Attribute condition fhir path
+- Attribute join fhir path
+- Attribute condition id and value fhir path
 
 ### Changed
 - Rename Conversion Template as Converter Template